Move initialization module
