+++ libs/bootstrap/__init__.py
@@ -4,4 +4,5 @@
 - `libs.bootstrap.configuration`: 初期設定関数
 - `libs.bootstrap.app_config`: アプリケーション設定関連
 - `libs.bootstrap.section`: セクション情報取り込み
+- `libs.bootstrap.initialization`: 各種初期化処理
 """
