+++ libs/bootstrap/configuration.py
@@ -18,9 +18,10 @@ import libs.commands.report.entry
 import libs.commands.results.entry
 import libs.global_value as g
 from integrations import factory
+from libs.bootstrap import initialization
 from libs.bootstrap.app_config import AppConfig
 from libs.commands.registry import member, team
-from libs.data import initialization, lookup
+from libs.data import lookup
 from libs.domain.datamodels import Args
 from libs.functions.compose import text_item
 from libs.types import ServiceType, StyleOptions
