+++ libs/commands/registry/team.py
@@ -6,8 +6,9 @@ import logging
 from typing import TYPE_CHECKING, TypedDict, cast
 
 import libs.global_value as g
+from libs.bootstrap import initialization
 from libs.bootstrap.section import BaseSection
-from libs.data import initialization, modify
+from libs.data import modify
 from libs.types import CommandType
 from libs.utils import dbutil, formatter, textutil, validator
